third_party/WebKit/LayoutTests/media/video-currentTime.html

<!DOCTYPE html>
<title>Test media current time values.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_equals(video.currentTime, 0);

    video.oncanplaythrough = t.step_func(function() {
        assert_equals(video.currentTime, 0);
        video.play();

        setTimeout(t.step_func_done(function() {
            assert_greater_than(video.currentTime, 0);
        }), 500);
    });

    video.src = "content/test.ogv";
});
</script>
